+++ NEWS.md
@@ -1,3 +1,5 @@
+# dfeshiny (development version)
+
 # dfeshiny 0.4.1
 
 * Fix bug in `external_link()` where visually hidden text was not visually